pyproject.toml
[tool.poetry]
name = "dact-pipeline"
version = "0.1.0"
description = "A data-driven test pipeline for AI chip/compiler testing"
authors = ["Gemini <gemini@example.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/example/dact-pipeline"
repository = "https://github.com/example/dact-pipeline"
documentation = "https://dact-pipeline.readthedocs.io"
keywords = ["testing", "ai", "chip", "compiler", "pipeline", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Hardware",
]
packages = [{include = "dact"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.2"
pydantic = "^2.11.7"
jinja2 = "^3.1.6"
typer = {extras = ["all"], version = "^0.16.0"}
rich = "^14.1.0"
pytest-html = "^4.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[tool.pytest.ini_options]
python_files = "tests/*.py"
python_classes = "*Tests"
python_functions = "test_*"
addopts = "-p dact.pytest_plugin"

[tool.poetry.scripts]
dact = "dact.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
